Backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
httpx>=0.23.0
python-dotenv>=1.0.0
pydantic-settings>=0.2.0
PyJWT>=2.8.0
email-validator>=1.3.0
